+++ packages/weapp-tailwindcss/src/constants.ts
@@ -13,6 +13,6 @@ export const WEAPP_TW_REQUIRED_NODE_VERSION = '18.17.0'
 
 export const DEFAULT_RUNTIME_PACKAGE_REPLACEMENTS = {
   'tailwind-merge': '@weapp-tailwindcss/merge',
-  'class-variance-authority': '@weapp-tailwindcss/merge/cva',
-  'tailwind-variants': '@weapp-tailwindcss/merge/variants',
+  'class-variance-authority': '@weapp-tailwindcss/cva',
+  'tailwind-variants': '@weapp-tailwindcss/variants',
 } as const